refactor(test): gather shared jestSilenceConsole into src/testSupport
The console-silencer is shared by the server and frontend jest setups, so
its only valid home is a common ancestor of both (src/). Move it under
src/testSupport/ to signal "shared test-only infra" rather than leaving it
loose at the src/ root.

Build behavior is unchanged: jestSilenceConsole and the jestSetupAfterEnv
files that import it stay excluded from the tsc build (ts-jest loads them
from source), so server output remains flat at dist/server/server.js. The
exclude glob is now **/jestSilenceConsole.ts so it tracks the file wherever
it lives.
