+++ src/frontend/jestSetupAfterEnv.ts
@@ -1,1 +1,1 @@
-import "../jestSilenceConsole";
+import "../testSupport/jestSilenceConsole";
