+++ tsconfig.json
@@ -16,11 +16,11 @@
         "**/*.test.ts",
         // Jest-only infra: transformed by ts-jest from source (see jest.config.js),
         // never consumed from dist/. Keeping these out of the build matters because
-        // jestSetupAfterEnv.ts imports ../jestSilenceConsole (a file at src/, outside
+        // jestSetupAfterEnv.ts imports ../testSupport/jestSilenceConsole (outside
         // src/server) — pulling it into the server program would drag the inferred
         // rootDir up to src/ and nest every output (dist/server/server/server.js).
         "**/jestGlobalSetup.ts",
         "**/jestSetupAfterEnv.ts",
-        "src/jestSilenceConsole.ts"
+        "**/jestSilenceConsole.ts"
     ]
 }
